requirements.txt
playwright
python-dotenv
django
djangorestframework
django-cors-headers
celery
redis
django-environ
